+++ test/test-3.jl
@@ -11,11 +11,6 @@ function main()
     carpet(400)
     setline(0.2)
 
-    mygfunction(vertices, faces, labels) = begin
-        randomhue()
-        Thebes.simplegfunction(vertices, faces, labels, action=:fill)
-    end
-
     for k in 1:4:20
         for i in 1:20
             object = make(Pyramid)
@@ -27,7 +22,7 @@ function main()
             rotateby!.(objects, 0, 0, a)
             sortfaces!(objects, eyepoint = eyepoint)
             # default is stroke at the moment
-            pin.(objects, gfunction = mygfunction)
+            pin.(objects)
         end
     end
     finish()